# [#25725] You can delete the default template style. Thanks Elin.
git-svn-id: http://joomlacode.org/svn/joomla/development/trunk@21186 6f6e1ebd-4c2b-0410-823f-f34bde69bce9

+++ administrator/components/com_templates/models/style.php
@@ -80,6 +80,11 @@ class TemplatesModelStyle extends JModelAdmin
 				if (!$user->authorise('core.delete', 'com_templates')) {
 					throw new Exception(JText::_('JERROR_CORE_DELETE_NOT_PERMITTED'));
 				}
+				// You should not delete a default style
+				if ($table->home != '0'){
+					JError::raiseWarning(SOME_ERROR_NUMBER,Jtext::_('COM_TEMPLATES_STYLE_CANNOT_DELETE_DEFAULT_STYLE'));
+					return false;
+				}
 
 				if (!$table->delete($pk)) {
 					$this->setError($table->getError());
